Fixed a bug that could generate exceptions when a new comment is added to a page

+++ meshcms3/src/java/org/meshcms/core/WebUtils.java
@@ -635,7 +635,7 @@ public final class WebUtils {
     } else if (cacheType == Configuration.ON_DISK_CACHE) {
       File cacheFile = getCacheFile(webSite, siteMap, pagePath);
 
-      if (cacheFile.exists()) {
+      if (cacheFile != null && cacheFile.exists()) {
         cacheFile.delete();
       }
     }
